refactor: extract suggested action context
