+++ src/lib.rs
@@ -1,3 +1,4 @@
+mod action_context;
 mod anchored_fs;
 pub mod app;
 mod bootstrap;
